pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "licitaciones-etl"
version = "0.1.0"
description = "ETL microservice CLI for licitaciones-espana (schemas, init-db, check-connection)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.28.0",
]

[project.scripts]
licitia-etl = "etl.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["etl*"]
